Implements a gids allocation schema based on the p-types generating tree model of M. Reimann et al, 2019 [BBPP82-60]
* The new algorithm is offered as an alternative to the previous one,
which is now referred to as the 'greedy' algorithm.
* The new algorithm, referred to as 'random', is selected by default.
* This stochastic algorithm relies on a p-types generating tree which is
built out of the data provided by the recipe. This generating tree
is the tree model defined in
"A null model of the mouse whole-neocortex micro-connectome"
by M. Reimann et al.
* The data used to build the tree are the innervation probabilities
of the target regions (frations in the recipe) and the interaction matrix, also
called matrix of increases in conditional innervation probabilities.
* If the data have been derived from an actual tree model, the re-constructed tree
generates p-types with the exact same probability distribution as the orginal tree.
* We do not fully check that the input data originate from a tree model.
We only ensure that the innervation probabilities and the interaction matrix are
compatible in a weak sense: 1.0 / I_S(A, B) >= max(P(S --> A), P(S -->B))
is guaranteed for every target regions A, B before building the tree.
* The random generation of p-types could be made faster if need be,
by pre-computing large amounts of random numbers beforehand.
(Currently computed on the fly.)

Change-Id: I80b41b405fac55565e922a87008486d7df2400ec